Revert "recycler: AccountListFragment enable recycler view"
This reverts commit 93f6dfe057547bc46f528d22cded34f850ffff8f.

I knew how to do it, but I still don't have a perfect solution to
bind ClickListener and LongClickListener to each rows, which is very
easy on ListView.

Besides that, I realized I don't have strong reason to replace ListView.

So just keep it as-is, until I got a reason.